refactor(agent-tars-web-ui): decouple message input and actionbar
